+++ app/src/main/java/fr/nelfdesign/oc_realestatemanager_p_9/propertylist/PhotoListViewModel.kt
@@ -17,7 +17,7 @@ class PhotoListViewModel(private val executor: Executor) : ViewModel() {
 
     val photos : LiveData<List<Photo>> = photoRepository.photos
 
-    fun getPhotoToDisplay(propertyId : Int) : LiveData<List<Photo>> = photoRepository.getPhotoToDisplay(propertyId)
+    fun getPhotoToDisplay(propertyId : Long) : LiveData<List<Photo>> = photoRepository.getPhotoToDisplay(propertyId)
 
     fun insertPhotos(photos : List<Photo>){
         executor.execute{
